refactor: move phase space factors to separate module
