+++ tests/sympy/test_math.py
@@ -3,7 +3,7 @@ import numpy as np
 import pytest
 import sympy as sp
 
-from ampform.dynamics import ComplexSqrt
+from ampform.sympy.math import ComplexSqrt
 
 
 class TestComplexSqrt:
